+++ src/banner/FPSWebPartClass/IThisFPSWebPartClass.ts
@@ -1,5 +1,6 @@
 
 import { DisplayMode } from '@microsoft/sp-core-library';
+import { IFPSEnviro } from '../../common/interfaces/fps/IFPSEnviro';
 import { WebPartContextCopy_15_2 } from '../../common/interfaces/indexes';
 import { ISitePreConfigProps } from '../../common/PropPaneHelp/IPreConfig';
 import { IRepoLinks } from '../../components/atoms/Links/CreateLinks';
@@ -42,12 +43,25 @@ export interface IThisFPSWebPartClass {
   _modifyBannerStyle: boolean;
   _keysToShow: ILoadPerformanceOps[]; // Performance keys to show in the propPaneHelp
 
+  //Disabling any of these removes from property pane, may not actually block importing though
+  _allowQuickLaunchHide: boolean; //Allows user to hide quick launch via property pane
+  _allowPageHeaderHide: boolean; //Allows user to hide Page Header (Title block) via property pane
+  _allowToolBarHide: boolean; //Allows user to hide Toolbar (like edit page button) via property pane
+  _allowAllSectWidth: boolean; //Allows user to hide Adjust all sections max width via property pane
+  _allowShowSearch: boolean; //Adds 'Show Search' Toggle to property pane, only needed if web part needs it.
+  _allowBeAUser: boolean; //Allows page editors to 'BeAUser' and adds to property pane where needed.
+  _allowFeedback: boolean; //Allows page editors to add 'Feedback email' in banner.
+  _allowExpando: boolean;  //Allows page editors the modify Expandoramic settings
+  _allowSiteThemeChoice: boolean; //Allows page editors the choice to inherit SiteTheme as a Theme choice
+  _allowEasyPages: boolean; //Allows page editors use EasyPages and EasyIcons
+
   /**
    * These are updated later in the code
    */
   _performance: ILoadPerformance;
   _sitePresets: ISitePreConfigProps;
   _FPSUser: IFPSUser;
+  _FPSEnviro: IFPSEnviro;
 
   /**
    * These are preset and should be managed by the code... do not change in main webpart class
@@ -76,7 +90,7 @@ export interface IThisFPSWebPartClass {
   domElement: HTMLElement;
   properties: IMinWPBannerProps;
   onPropertyPaneConfigurationStart: any;
-
+  onPropertyPaneFieldChanged: any;