+++ src/banner/bannerX/FullPanel.tsx
@@ -109,11 +109,14 @@ export function getFullPanel (  bannerProps: IWebpartBannerProps, selectedKey: s
 
       if ( showMedical === true ) {
         let medicalElements : any = [];
+        if ( thisWindow.FPSEnviro ) {
+          medicalElements.push( <ReactJson src={ thisWindow.FPSEnviro } name={ 'FPSEnviro' } collapsed={ false } displayDataTypes={ true } displayObjectSize={ true } enableClipboard={ false } style={{ padding: '10px 0px' }}/> );
+        }
         if ( thisWindow.FPSUser ) {
           medicalElements.push( <ReactJson src={ thisWindow.FPSUser } name={ 'FPSUser' } collapsed={ true } displayDataTypes={ true } displayObjectSize={ true } enableClipboard={ false } style={{ padding: '10px 0px' }}/> );
         }
-        if ( thisWindow.FPSUser ) {
-          medicalElements.push( <ReactJson src={ thisWindow.FPSOptions } name={ 'FPSOptions' } collapsed={ false } displayDataTypes={ true } displayObjectSize={ true } enableClipboard={ false } style={{ padding: '10px 0px' }}/> );
+        if ( thisWindow.FPSOptions ) {
+          medicalElements.push( <ReactJson src={ thisWindow.FPSOptions } name={ 'FPSOptions' } collapsed={ true } displayDataTypes={ true } displayObjectSize={ true } enableClipboard={ false } style={{ padding: '10px 0px' }}/> );
         }
         content= <div id="MedicalPanel" style={{paddingTop: '20px'}}>
           { medicalElements }