+++ src/banner/propPane/FPSOptionsGroupBasic.ts
@@ -3,7 +3,10 @@ import {
   PropertyPaneToggle, PropertyPaneSlider
 } from '@microsoft/sp-property-pane';
 import { IFPSBasicToggleSetting } from '../../common/interfaces/fps/IFPSBasicToggleSetting';
-
+import { check4Gulp } from '@mikezimm/fps-pnp2/lib/services/sp/CheckGulping';
+import { IFPSWindow } from '../../common/interfaces/fps/Window';
+import { IPageLayoutType } from '../../common/interfaces/indexes/Layout';
+import { IThisFPSWebPartClass } from '../FPSWebPartClass/IThisFPSWebPartClass';
 
 /**
  * This is the second version which is more simple (toggles and sliders)
@@ -13,24 +16,30 @@ import { IFPSBasicToggleSetting } from '../../common/interfaces/fps/IFPSBasicTog
  * @param containerMaxWidth
  */
 
+//thisWPClass: IThisFPSWebPartClass
+// export function FPSOptionsGroupBasic(showSearch: boolean, quickLaunchHide: boolean, pageHeaderHide: boolean, allSectWidth: IFPSBasicToggleSetting, 
+//     allSectionMaxWidthEnable: any, allSectMargin: IFPSBasicToggleSetting, allSectionMarginEnable: any, toolBarHide: IFPSBasicToggleSetting) {
 
-export function FPSOptionsGroupBasic(showSearch: boolean, quickLaunchHide: boolean, pageHeaderHide: boolean, allSectWidth: IFPSBasicToggleSetting, allSectionMaxWidthEnable: any, allSectMargin: IFPSBasicToggleSetting, allSectionMarginEnable: any, toolBarHide: IFPSBasicToggleSetting) {
+export function FPSOptionsGroupBasic(thisWPClass: IThisFPSWebPartClass) {
 
+  const isSinglePageApp = thisWPClass._FPSEnviro.pageLayout === 'SingleWebPartAppPageLayout' ? true : false;
+
+  if ( isSinglePageApp === true ) return 
   let fields: any[] = [];
-  if (showSearch === true) {
+  if (thisWPClass._allowShowSearch === true) {
     fields.push(
       PropertyPaneToggle('searchShow', {
         label: 'Show search bar by default', offText: 'Hide', onText: 'Show',
       }));
   }
-  if (quickLaunchHide === true) {
+  if (thisWPClass._allowQuickLaunchHide === true) {
     fields.push(
       PropertyPaneToggle('quickLaunchHide', {
         label: 'Hide quick launch - may be seen briefly', offText: 'Ignore', onText: 'Hidden',
       }));
   }
 
-  if (pageHeaderHide === true) {
+  if (thisWPClass._allowPageHeaderHide === true) {
     fields.push(
       PropertyPaneToggle('pageHeaderHide', {
         label: 'Hide Page Header - may be seen briefly',
@@ -39,7 +48,7 @@ export function FPSOptionsGroupBasic(showSearch: boolean, quickLaunchHide: boole
       })
     );
   }
-  if (allSectWidth !== 'skip') {
+  if (thisWPClass._allowAllSectWidth !== false ) {
     fields.push(
       PropertyPaneToggle('allSectionMaxWidthEnable', {
         label: 'All Sections Max Width',
@@ -50,14 +59,14 @@ export function FPSOptionsGroupBasic(showSearch: boolean, quickLaunchHide: boole
     fields.push(
       PropertyPaneSlider('allSectionMaxWidth', {
         label: 'Max width of all sections',
-        disabled: allSectionMaxWidthEnable === true ? false : true,
+        disabled: thisWPClass.properties.allSectionMaxWidthEnable === true ? false : true,
         min: 1200,
         max: 3200,
         step: 100,
       })
     );
   }
-  if (allSectMargin !== 'skip') {
+  if (thisWPClass._allowAllSectWidth !== false) {
     fields.push(
       PropertyPaneToggle('allSectionMarginEnable', {
         label: 'All Sections Margin',
@@ -68,14 +77,14 @@ export function FPSOptionsGroupBasic(showSearch: boolean, quickLaunchHide: boole
     fields.push(
       PropertyPaneSlider('allSectionMargin', {
         label: 'Top and Bottom Margin',
-        disabled: allSectionMarginEnable === true ? false : true,
+        disabled: thisWPClass.properties.allSectionMarginEnable === true ? false : true,
         min: 0,
         max: 100,
         step: 2,
       })
     );
   }
-  if (toolBarHide === true) {
+  if (thisWPClass._allowToolBarHide === true) {
     fields.push(
       PropertyPaneToggle('toolBarHide', {
         label: 'Hide Toolbar - while viewing',