+++ src/banner/propPane/WebPartInfoGroup.ts
@@ -9,7 +9,7 @@ import { fpsLogo326 } from '../../components/atoms/SVGIcons/fpsLogo326';
 import { IRepoLinks } from '../../components/atoms/Links/CreateLinks';
 
 /**
- * 
+ *
  * @param gitRepoLinks 
  * @param shortDescription 
  * @param PropertyPaneWebPartInformation import { PropertyPaneWebPartInformation } from '@pnp/spfx-property-controls/lib/PropertyPaneWebPartInformation';