+++ src/common/interfaces/fps/Window.ts
@@ -1,11 +1,13 @@
 import { IFPSWindowProps } from "../../../banner/features/FPSDOM/Interfaces";
 import { IFPSUser } from "../../../logic/indexes"
-
+import { IFPSEnviro } from "./IFPSEnviro";
 
 export interface IFPSWindow {
 
   FPSUser: IFPSUser;
 
   FPSOptions: IFPSWindowProps;
+
+  FPSEnviro: IFPSEnviro;
 
 }